Added ant simulations for moving routine cars on the city map
The traffic server has been modified so that now ant simulation can
also be used to move routine cars. The class AntCar is derived from
class Car and overrides the virtual method nextSmarterEdge() where I
have been experimenting with several settings:

TrafficType::ANT

    the routine cars select the edge that was chosen the most number
of times out of the possible edges in the previous steps.

TrafficType::ANT_RND

    the next edge will be chosen with a probability that increases
with the number of past selections of the given edge.

TrafficType::ANT_RERND

    it is similar to the previous one, but the sum of the previous
probabilities of all possible edge is only equal to 3/4.

The actual results can be seen on the attached screenshots. But
routine cars still cannot exhibit ant-like behavior... I would like to
provide a solution that is more similar to patterns that can be seen
in the ant simulations of
http://javacska.lib.unideb.hu/programkak/hangyak.html or
http://javacska.lib.unideb.hu/hangyak/